Add new module for C++ vectors of diploids.
