+++ fwdpy11/src/fwdpy11_types.cc
@@ -41,39 +41,6 @@ PYBIND11_MAKE_OPAQUE(
     std::vector<double>); // for generalmut_vec::s and generalmut_vec::h
 
 
-struct diploid_traits
-{
-    double g, e, w;
-};
-
-inline diploid_traits
-make_diploid_traits(const fwdpy11::diploid_t& dip)
-{
-    diploid_traits d;
-    d.g = dip.g;
-    d.e = dip.e;
-    d.w = dip.w;
-    return d;
-}
-
-struct diploid_gametes
-{
-    std::size_t locus, first, second;
-};
-
-inline diploid_gametes
-make_diploid_gametes(const fwdpy11::diploid_t& dip, const std::size_t locus)
-{
-    diploid_gametes d;
-    d.locus = locus;
-    d.first = dip.first;
-    d.second = dip.second;
-    return d;
-}
-
-PYBIND11_MAKE_OPAQUE(std::vector<diploid_traits>);
-PYBIND11_MAKE_OPAQUE(std::vector<diploid_gametes>);
-
 namespace
 {
     static const auto MCOUNTS_DOCSTRING = R"delim(
@@ -173,292 +140,6 @@ PYBIND11_MODULE(fwdpy11_types, m)
         .def("__eq__", [](const fwdpy11::diploid_t& a,
                           const fwdpy11::diploid_t& b) { return a == b; });
 
-    py::bind_vector<fwdpy11::dipvector_t>(
-        m, "DiploidContainer", py::module_local(false),
-        "C++ representation of a list of "
-        ":class:`fwdpy11.fwdpy11_types."
-        "SingleLocusDiploid`.  Typically, access will be read-only.")
-        .def("trait_array",
-             [](const fwdpy11::dipvector_t& diploids) {
-                 std::vector<diploid_traits> rv;
-                 rv.reserve(diploids.size());
-                 for (auto&& dip : diploids)
-                     {
-                         rv.push_back(make_diploid_traits(dip));
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipTraits`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("trait_array",
-             [](const fwdpy11::dipvector_t& diploids,
-                py::array_t<std::size_t> individuals) {
-                 auto r = individuals.unchecked<1>();
-                 std::vector<diploid_traits> rv;
-                 rv.reserve(r.shape(0));
-                 for (decltype(r.shape(0)) i = 0; i < r.shape(0); ++i)
-                     {
-                         // range-check here
-                         auto&& dip = diploids.at(r(i));
-                         rv.push_back(make_diploid_traits(dip));
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipTraits`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("trait_array",
-             [](const fwdpy11::dipvector_t& diploids, py::slice slice) {
-                 size_t start, stop, step, slicelength;
-
-                 if (!slice.compute(diploids.size(), &start, &stop, &step,
-                                    &slicelength))
-                     throw py::error_already_set();
-
-                 std::vector<diploid_traits> rv;
-                 rv.reserve(slicelength);
-                 for (size_t i = 0; i < slicelength; ++i)
-                     {
-                         auto&& dip = diploids.at(start);
-                         rv.push_back(make_diploid_traits(dip));
-                         start += step;
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipTraits`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("key_array",
-             [](const fwdpy11::dipvector_t& diploids) {
-                 std::vector<diploid_gametes> rv;
-                 rv.reserve(diploids.size());
-                 for (auto&& dip : diploids)
-                     {
-                         rv.push_back(make_diploid_gametes(dip, 0));
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipGametes`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("key_array",
-             [](const fwdpy11::dipvector_t& diploids,
-                py::array_t<std::size_t> individuals) {
-                 auto r = individuals.unchecked<1>();
-                 std::vector<diploid_gametes> rv;
-                 rv.reserve(r.shape(0));
-                 for (decltype(r.shape(0)) i = 0; i < r.shape(0); ++i)
-                     {
-                         rv.push_back(make_diploid_gametes(diploids.at(i), 0));
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipGametes`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("key_array",
-             [](const fwdpy11::dipvector_t& diploids, py::slice slice) {
-                 size_t start, stop, step, slicelength;
-
-                 if (!slice.compute(diploids.size(), &start, &stop, &step,
-                                    &slicelength))
-                     throw py::error_already_set();
-
-                 std::vector<diploid_gametes> rv;
-                 rv.reserve(slicelength);
-                 for (size_t i = 0; i < slicelength; ++i)
-                     {
-                         rv.push_back(
-                             make_diploid_gametes(diploids.at(start), 0));
-                         start += step;
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipGametes`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def(py::pickle(
-            [](const std::vector<fwdpy11::diploid_t>& v) -> py::list {
-                py::list rv;
-                for (auto&& vi : v)
-                    {
-                        rv.append(vi);
-                    }
-                return rv;
-            },
-            [](py::list l) {
-                std::vector<fwdpy11::diploid_t> rv;
-                for (auto&& i : l)
-                    {
-                        rv.push_back(i.cast<fwdpy11::diploid_t>());
-                    }
-                return rv;
-            }));
-
-    py::bind_vector<std::vector<fwdpy11::dipvector_t>>(
-        m, "VecDiploidContainer", py::module_local(false),
-        "Vector of "
-        ":class:`fwdpy11.fwdpy11_types."
-        "SingleLocusDiploid`.")
-        .def("trait_array",
-             [](const std::vector<fwdpy11::dipvector_t>& diploids) {
-                 std::vector<diploid_traits> rv;
-                 rv.reserve(diploids.size());
-                 for (auto&& dip : diploids)
-                     {
-                         rv.emplace_back(make_diploid_traits(dip.at(0)));
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipTraits`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("key_array",
-             [](const std::vector<fwdpy11::dipvector_t>& diploids) {
-                 std::vector<diploid_gametes> rv;
-                 std::size_t locus;
-                 for (auto&& dip : diploids)
-                     {
-                         locus = 0;
-                         for (auto&& di : dip)
-                             {
-                                 rv.emplace_back(
-                                     make_diploid_gametes(di, locus++));
-                             }
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipGametes`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("trait_array",
-             [](const std::vector<fwdpy11::dipvector_t>& diploids,
-                py::array_t<std::size_t> individuals) {
-                 auto r = individuals.unchecked<1>();
-                 std::vector<diploid_traits> rv;
-                 rv.reserve(r.shape(0));
-                 for (decltype(r.shape(0)) i = 0; i < r.shape(0); ++i)
-                     {
-                         // range-check here
-                         auto&& dip = diploids.at(r(i)).at(0);
-                         rv.push_back(make_diploid_traits(dip));
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipTraits`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("key_array",
-             [](const std::vector<fwdpy11::dipvector_t>& diploids,
-                py::array_t<std::size_t> individuals) {
-                 auto r = individuals.unchecked<1>();
-                 std::vector<diploid_gametes> rv;
-                 rv.reserve(r.shape(0));
-                 std::size_t locus;
-                 for (decltype(r.shape(0)) i = 0; i < r.shape(0); ++i)
-                     {
-                         auto&& dip = diploids.at(r(i));
-                         locus = 0;
-                         for (auto&& di : dip)
-                             {
-                                 rv.emplace_back(
-                                     make_diploid_gametes(di, locus++));
-                             }
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipGametes`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("trait_array",
-             [](const std::vector<fwdpy11::dipvector_t>& diploids,
-                py::slice slice) {
-                 size_t start, stop, step, slicelength;
-
-                 if (!slice.compute(diploids.size(), &start, &stop, &step,
-                                    &slicelength))
-                     throw py::error_already_set();
-
-                 std::vector<diploid_traits> rv;
-                 rv.reserve(slicelength);
-                 for (size_t i = 0; i < slicelength; ++i)
-                     {
-                         auto&& dip = diploids.at(start).at(0);
-                         rv.push_back(make_diploid_traits(dip));
-                         start += step;
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipTraits`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def("key_array",
-             [](const std::vector<fwdpy11::dipvector_t>& diploids,
-                py::slice slice) {
-                 size_t start, stop, step, slicelength;
-
-                 if (!slice.compute(diploids.size(), &start, &stop, &step,
-                                    &slicelength))
-                     throw py::error_already_set();
-
-                 std::vector<diploid_gametes> rv;
-                 rv.reserve(slicelength);
-                 std::size_t locus;
-                 for (size_t i = 0; i < slicelength; ++i)
-                     {
-                         auto&& dip = diploids.at(start).at(0);
-                         locus = 0;
-                         rv.push_back(make_diploid_gametes(dip, locus++));
-                         start += step;
-                     }
-                 return rv;
-             },
-             R"delim(
-			 :rtype: :class:`fwdpy11.fwdpy11_types.VecDipGametes`
-			 
-			 .. versionadded:: 0.1.2
-			 )delim")
-        .def(py::pickle(
-            [](const std::vector<fwdpy11::dipvector_t>& diploids) {
-                py::list rv;
-                for (auto&& i : diploids)
-                    {
-                        rv.append(i);
-                    };
-                return rv;
-            },
-            [](py::list l) {
-                std::vector<fwdpy11::dipvector_t> rv;
-                for (auto&& i : l)
-                    {
-                        rv.push_back(i.cast<fwdpy11::dipvector_t>());
-                    }
-                return rv;
-            }));
-
     py::bind_vector<std::vector<KTfwd::uint_t>>(
         m, "VectorUint32", "Vector of unsigned 32-bit integers.",
         py::buffer_protocol(), py::module_local(false))
@@ -480,27 +161,6 @@ PYBIND11_MODULE(fwdpy11_types, m)
                 return rv;
             }));
 
-
-    PYBIND11_NUMPY_DTYPE(diploid_traits, g, e, w);
-    PYBIND11_NUMPY_DTYPE(diploid_gametes, locus, first, second);
-
-    py::bind_vector<std::vector<diploid_traits>>(
-        m, "VecDipTraits", py::buffer_protocol(), py::module_local(false),
-        R"delim(
-        Vector of the g,e,w data fields in a "
-        ":class:`fwdpy11.fwdpp_types.SingleLocusDiploid`.
-
-        .. versionadded: 0.1.2
-        )delim");
-
-    py::bind_vector<std::vector<diploid_gametes>>(
-        m, "VecDipGametes", py::buffer_protocol(), py::module_local(false),
-        R"delim(
-        Vector of the locus and gamete index data fields in a "
-        ":class:`fwdpy11.fwdpp_types.SingleLocusDiploid`.
-
-        .. versionadded: 0.1.2
-        )delim");
 
     // expose the base classes for population types
     py::class_<fwdpp_popgenmut_base>(m, "SlocusPopMutationBase");
